requirements.txt
beautifulsoup4
numpy
opencv_python
Pillow
PyAutoGUI
pynput
Requests
